requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
distlib==0.3.1
Django==3.0.8
filelock==3.0.12
idna==2.10
newsapi-python==0.2.6
pipenv==2020.6.2
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
virtualenv==20.0.25
virtualenv-clone==0.5.4
